requirements.txt
openai
tweepy
apscheduler
python-dotenv
requests
numpy
flask